third_party/WebKit/LayoutTests/editing/pasteboard/4944770-2.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p>This tests smart paste of a fragment that ends in a select element.
There should be spaces added before and after the inserted content.</p>
<div contenteditable="true" id="copy"><select><option>1</option><option>2</option><option>3</option></select></div>
<div contenteditable="true" id="sample">xx</div>
<div id="log"></div>
<script>
test(function() {
    if (window.internals)
        internals.settings.setEditingBehavior('win');

    var selection = window.getSelection();
    var sample= document.getElementById('sample');

    selection.selectAllChildren(document.getElementById('copy'));
    document.execCommand('copy');

    selection.collapse(sample.firstChild, 1);
    document.execCommand('paste');

    assert_equals(sample.innerHTML, 'x<select><option>1</option><option>2</option><option>3</option></select>x', 'innerHHTML');
    assert_true(selection.isCollapsed);
    assert_equals(selection.anchorNode, sample);
    assert_equals(selection.anchorOffset, 2);
});
</script>
